create MessageContentView, MessageView
