+++ js/templates/messagecontent.html
@@ -1,0 +1,2 @@
+<div id="mail-messages"></div>
+<div id="mail-message" class=" hidden-mobile"></div>
